Upgraded to Vaadin 14 because of IntegerField.

+++ src/main/java/hu/pte/ttk/vaadin/vaadindemo/car/entity/CarEntity.java
@@ -29,6 +29,6 @@ public class CarEntity extends CoreEntity {
 	private Integer carDoors;
 
 	@Column(name = "car_manufactured")
-	private Double carManufactured;
+	private Integer carManufactured;
 
 }
